+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This document follows the conventions laid out in [Keep a CHANGELOG](https://keepachangelog.com/en/1.0.0).
 
+## [Unreleased]()
+
+### Added
+
+- Added custom argument for the copyright's start year & update add_hook_changes ([#111](https://github.com/ansys/pre-commit-hooks/pull/111))
+
 ## [0.2.4](https://github.com/ansys/pre-commit-hooks/releases/tag/v0.2.4) - January 4 2024
 
 ### Fixed